# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cash Bank apis",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "https://cashbank-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "get all transactions",
                "operationId": "11e1653062ed9cbaafdf09bacf28383b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "any"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "data base column name"
                        }
                    },
                    {
                        "name": "limited",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "store new transaction",
                "operationId": "722021fd72f5d9e61aaec7a7f04d34d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/transactions/{transaction}": {
            "get": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "show one transaction",
                "operationId": "c94e6a6347327949ca06be6f99a905fc",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/get-amount-received": {
            "post": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "get the amount received",
                "operationId": "75f5b71f37fc757a3570e8ba32bb6e75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount_sent": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/transactions/get-amount-sent": {
            "post": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "get the amount sent",
                "operationId": "512e90e80a7a57f16e0b7b9e8beb804a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount_received": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/transactions/{transaction}/confirm-transform": {
            "post": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "confirm transformation",
                "operationId": "86748b37122cfcfab8de51b30cddaf3e",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/transactions/{transaction}/cancel-transform": {
            "post": {
                "tags": [
                    "App",
                    "App - Transaction"
                ],
                "summary": "cancel transformation",
                "operationId": "131b34da486e23e42c3a922c77592720",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateTransactionRequest": {
                "properties": {
                    "amount_sent": {
                        "type": "number",
                        "format": "double"
                    },
                    "amount_received": {
                        "type": "number",
                        "format": "double"
                    },
                    "from_wallet_id": {
                        "type": "integer"
                    },
                    "to_wallet_id": {
                        "type": "integer"
                    },
                    "from_wallet_number": {
                        "type": "string"
                    },
                    "to_wallet_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TransactionResource": {
                "title": "TransactionResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}